requirements.txt
# core
python-dotenv>=1.0
requests>=2.28
pandas>=2.0
matplotlib>=3.7

# YouTube transcript / download
youtube-transcript-api>=0.6.1
pytube>=12.1.0

# LLM SDKs (placeholders — you'll add real API keys)
openai>=1.0.0
google-generativeai>=0.1.0

# optional dev
black
pytest
